refactor(core): deprecate initiateRecovery in xlm (1 of 2)
Add getStellarKeys to initiate. Next step we can inline
initiateRecovery.

Issue: BG-34381
