+++ modules/core/src/v2/recovery/initiate.ts
@@ -4,6 +4,7 @@
 
 import * as utxolib from '@bitgo/utxo-lib';
 import { BitGo } from '../../bitgo';
+import * as stellar from 'stellar-sdk';
 
 interface ValidateKeyOptions {
   key: string;
@@ -83,6 +84,49 @@ export function getBip32Keys(
     if (requireBitGoXpub) {
       throw new Error('Failed to parse bitgo xpub!');
     }
+  }
+
+  return keys;
+}
+
+export function getStellarKeys(bitgo: BitGo, params: InitiateRecoveryOptions): stellar.Keypair[] {
+  const keys: stellar.Keypair[] = [];
+  let userKey = params.userKey;
+  let backupKey = params.backupKey;
+
+  // Stellar's Ed25519 public keys start with a G, while private keys start with an S
+  const isKrsRecovery = backupKey.startsWith('G') && !userKey.startsWith('G');
+  const isUnsignedSweep = backupKey.startsWith('G') && userKey.startsWith('G');
+
+  try {
+    if (!userKey.startsWith('S') && !userKey.startsWith('G')) {
+      userKey = bitgo.decrypt({
+        input: userKey,
+        password: params.walletPassphrase,
+      });
+    }
+
+    const userKeyPair = isUnsignedSweep ? stellar.Keypair.fromPublicKey(userKey) : stellar.Keypair.fromSecret(userKey);
+    keys.push(userKeyPair);
+  } catch (e) {
+    throw new Error('Failed to decrypt user key with passcode - try again!');
+  }
+
+  try {
+    if (!backupKey.startsWith('S') && !isKrsRecovery && !isUnsignedSweep) {
+      backupKey = bitgo.decrypt({
+        input: backupKey,
+        password: params.walletPassphrase,
+      });
+    }
+
+    if (isKrsRecovery || isUnsignedSweep) {
+      keys.push(stellar.Keypair.fromPublicKey(backupKey));
+    } else {
+      keys.push(stellar.Keypair.fromSecret(backupKey));
+    }
+  } catch (e) {
+    throw new Error('Failed to decrypt backup key with passcode - try again!');
   }
 
   return keys;
